bug 1529342: rename ProcessorPipeline to Pipeline
This reduces some redundancy in the naming and makes it easier to type.
